+++ main.py
@@ -1,1 +1,1 @@
-import controller
+import controllers.controller as controller
